+++ tests/testthat/test-sigchld.R
@@ -112,3 +112,80 @@ test_that("SIGCHLD handler", {
 
   expect_true(px$get_exit_status() %in% c(0L, NA_integer_))
 })
+
+test_that("Notify old signal handler", {
+  skip_on_cran()
+  skip_other_platforms("unix")
+
+  code <- substitute({
+    # Create cluster, check that it works
+    cl <- parallel::makeForkCluster(2)
+    parallel::mclapply(1:2, function(x) x)
+
+    # Run a parallel background job
+    job <- parallel::mcparallel(Sys.sleep(.5))
+
+    # Start processx process, it will overwrite the signal handler
+    processx::run("true")
+
+    # Wait for parallel job to finish
+    parallel::mccollect(job)
+  })
+
+  script <- tempfile(pattern = "processx-test-", fileext = ".R")
+  on.exit(unlink(script), add = TRUE)
+  cat(deparse(code), sep = "\n", file = script)
+
+  env <- c(callr::rcmd_safe_env(), PROCESSX_NOTIFY_OLD_SIGCHLD = "true")
+  ret <- callr::rscript(
+    script,
+    env = env,
+    fail_on_status = FALSE,
+    show = FALSE,
+    timeout = 5
+  )
+
+  # parallel sends a message to stderr, complaining about unable to
+  # to terminate some child processes. That should not happen any more.
+  expect_equal(ret$stderr, "")
+})
+
+test_that("it is ok if parallel has no active cluster", {
+  skip_on_cran()
+  skip_other_platforms("unix")
+
+  code <- substitute({
+    cl <- parallel::makeForkCluster(2)
+    parallel::mclapply(1:2, function(x) x)
+
+    job <- parallel::mcparallel(Sys.sleep(.5))
+    processx::run("true")
+    parallel::mccollect(job)
+
+    # stop cluster, verify that we don't have subprocesses
+    parallel::stopCluster(cl)
+    print(ps::ps_children(ps::ps_handle()))
+
+    # try to run sg, this still calls the old sigchld handler
+    for (i in 1:5) processx::run("true")
+
+    # No cluster, just to clarify
+    print(parallel::getDefaultCluster())
+  })
+
+  script <- tempfile(pattern = "processx-test-", fileext = ".R")
+  on.exit(unlink(script), add = TRUE)
+  cat(deparse(code), sep = "\n", file = script)
+
+  env <- c(callr::rcmd_safe_env(), PROCESSX_NOTIFY_OLD_SIGCHLD = "true")
+  ret <- callr::rscript(
+    script,
+    env = env,
+    fail_on_status = FALSE,
+    show = FALSE
+  )
+
+  expect_equal(ret$status, 0)
+  expect_match(ret$stdout, "list()")
+  expect_match(ret$stdout, "NULL")
+})